Animate variable chooser transitions
