+++ src/components/Form/Fields/index.js
@@ -1,0 +1,2 @@
+export { default as Tag } from './Tag';
+export { default as VariableChooser } from './VariableChooser';
