+++ src/components/StageEditor/sections/NameGeneratorPrompts/NameGeneratorPrompt.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'redux-form';
-import { SeamlessTextInput, VariableChooser } from '../../../Form';
+import { SeamlessTextInput } from '../../../Form';
+import { VariableChooser } from '../../../Form/Fields';
 
 const NameGeneratorPrompt = ({ fieldId, form, variableRegistry }) => (
   <div className="prompt">
